remake the piece structure
I removed the individual classes for each type of piece and rewrote the
piece class. I also created the board array and class to manage the
pieces and their operations. The piece class is updated to hold
information about the pieces' color, type, and location on the board. I
created two linked stacks in the board class for taken pieces on either
side.